🎨 design(): fixed sidebar height
The sidebar was tied to the height of the page. If the content is long vertically, the bottom of the sidebar would be all the way down in the page.

I wrapped the nav in a div and fixed the position of the sidebar.

Added a border right on the sidebar